requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.13.0
scikit-learn==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
pandas==2.0.3